LLM/requirements.txt
beautifulsoup4==4.12.3
faiss-cpu==1.9.0
fastapi==0.112.4
Flask==3.1.0
httpcore==0.16.3
httpx==0.23.3
httpx-sse==0.4.0
ipython==8.26.0
json5==0.9.25
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
networkx==3.4.2
openai==1.54.5
python-dotenv==1.0.1
requests==2.32.3
scipy==1.14.1
sentence-transformers==3.3.1
streamlit==1.39.0
transformers==4.46.3
urllib3==2.2.2